// FootballManagement/football-ng/src/app/Model/team-in-tournament.ts
export class TeamInTournament {
  id: number;
  teamName: string;
  teamId: number;
  logoTeam: string;
  tournamentName: string;
  tournamentId: number;
  outJScore: number;
  homeJScore: number;
  totalScore: number;
  status: number;
  color: string;
  locationName: string;
  mW: number;
  mD: number;
  mL: number;
  gF: number;
  gA: number;
  diff: number;
  weekNo: string;
}